+++ cypress/integration/application.spec.ts
@@ -7,18 +7,37 @@ describe('/', () => {
     });
   });
 
-  context('when I am logged in', () => {
+  context('when I am logged in as BEIS user', () => {
     beforeEach(() => {
       cy.loginAuth0('admin');
       cy.visitAndCheckAccessibility('/admin');
     });
 
-    it('shows my name', () => {
+    it('shows my name and the BEIS organisation name', () => {
       cy.translate('app.welcome', { name: 'beis-rpr' }).then(
         (welcomeMessage) => {
           cy.get('body').should('contain', welcomeMessage);
         },
       );
+      cy.translate('app.beis').then((beis) => {
+        cy.get('span').should('contain', beis);
+      });
+    });
+  });
+
+  context('when I am logged in as a non BEIS user', () => {
+    beforeEach(() => {
+      cy.loginAuth0('orgadmin');
+      cy.visitAndCheckAccessibility('/admin');
+    });
+
+    it('shows my name and organisation', () => {
+      cy.translate('app.welcome', { name: 'Organisation Admin' }).then(
+        (welcomeMessage) => {
+          cy.get('body').should('contain', welcomeMessage);
+        },
+      );
+      cy.get('span').should('contain', 'Department for Education');
     });
   });
 });