+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Show user's name rather than username in welcome message
 - Display additional UK qualification data where the Profession itself is not a whole-UK Profession
 - Remove non-UK qualification data
+- Show user's organisation on dashboard
 
 ## [release-007] - 2022-03-04